+++ src/components/SideBar/SideBarForMobile.tsx
@@ -1,0 +1,5 @@
+import { memo } from 'react';
+
+export const SideBarForMobile = memo(function SideBarForMobile() {
+    return null;
+});
